refactor(daemon): split start workflow into recovery and launch services
Why:
- DaemonStartOperation に回復処理と起動補償処理が集中しており凝集が低かったため。

What Changed:
- Daemon.Start 名前空間に recovery/launch の専用サービスと契約を追加。
- DaemonStartOperation を分岐制御のオーケストレーション責務へ縮小。

Impact:
- start フローの責務境界が明確になり、テストと保守がしやすくなる。
