+++ src/Ucli/Daemon/Start/IDaemonStartLaunchService.cs
@@ -1,0 +1,17 @@
+using MackySoft.Ucli.UnityProject;
+
+namespace MackySoft.Ucli.Daemon.Start;
+
+/// <summary> Launches daemon process and applies startup-compensation workflow when start fails. </summary>
+internal interface IDaemonStartLaunchService
+{
+    /// <summary> Launches daemon lifecycle for the specified Unity project context. </summary>
+    /// <param name="unityProject"> The resolved Unity project context. </param>
+    /// <param name="timeout"> The daemon startup timeout. </param>
+    /// <param name="cancellationToken"> The cancellation token propagated by command execution. </param>
+    /// <returns> The daemon start result. </returns>
+    ValueTask<DaemonStartResult> Launch (
+        ResolvedUnityProjectContext unityProject,
+        TimeSpan timeout,
+        CancellationToken cancellationToken = default);
+}
